Remove more plugins
Removed the layout plugin and replaced it with a layout component.

Noticed and fixed the problem with graph objects not linking. Obvious
once you see it, but it can only work with core fields. Luckily, I can
work around that with a custom resolver. This will also mean it's much
easier to merge data from previews and lunar in the resolver.